+++ packages/terra-content-container/tests/wdio/content-container-spec.js
@@ -2,12 +2,12 @@ Terra.describeViewports('Content Container', ['medium'], () => {
   it('displays default Content Container', () => {
     browser.url('/raw/tests/terra-content-container/content-container/default-content-container');
 
-    Terra.validates.element('default');
+    Terra.validates.element('content container');
   });
 
   it('displays Content Container with fill', () => {
     browser.url('/raw/tests/terra-content-container/content-container/fill-content-container');
 
-    Terra.validates.element('fill');
+    Terra.validates.element('content container fill');
   });
 });