+++ packages/terra-text/tests/wdio/text-spec.js
@@ -1,56 +1,56 @@
 Terra.describeViewports('Text', ['huge'], () => {
-  it('displays Default text', () => {
+  it('should display default text', () => {
     browser.url('/raw/tests/terra-text/text/default-text');
-    Terra.validates.element('default');
+    Terra.validates.element('default text');
   });
 
-  it('displays text with Color', () => {
+  it('should display text with Color', () => {
     browser.url('/raw/tests/terra-text/text/text-color');
-    Terra.validates.element('color');
+    Terra.validates.element('color text');
   });
 
-  it('displays text with Italics', () => {
+  it('should display text with italics', () => {
     browser.url('/raw/tests/terra-text/text/text-italics');
-    Terra.validates.element('italics');
+    Terra.validates.element('italic text');
   });
 
-  it('displays Font Size Group 1', () => {
+  it('should display font size group 1', () => {
     browser.url('/raw/tests/terra-text/text/text-font-size-1');
     Terra.validates.element('font size group 1');
   });
 
-  it('Font Size Group 2', () => {
+  it('should display font size group 2', () => {
     browser.url('/raw/tests/terra-text/text/text-font-size-2');
     Terra.validates.element('font size group 2');
   });
 
-  it('displays Font Size Group 3', () => {
+  it('should display font size group 3', () => {
     browser.url('/raw/tests/terra-text/text/text-font-size-3');
     Terra.validates.element('font size group 3');
   });
 
-  it('displays Visually Hidden text', () => {
+  it('should display visually hidden text', () => {
     browser.url('/raw/tests/terra-text/text/text-visually-hidden');
-    Terra.validates.element('visually hidden');
+    Terra.validates.element('visually hidden text');
   });
 
-  it('displays with Visual Props', () => {
+  it('should display with visual props', () => {
     browser.url('/raw/tests/terra-text/text/text-visual-props');
     Terra.validates.element('visual props');
   });
 
-  it('displays with Weight', () => {
+  it('should display with weight', () => {
     browser.url('/raw/tests/terra-text/text/text-weight');
-    Terra.validates.element('weight');
+    Terra.validates.element('weight text');
   });
 
-  it('should calculate Line-height on text font size when parent has a different font size.', () => {
+  it('should calculate line-height on text font size when parent has a different font size.', () => {
     browser.url('/raw/tests/terra-text/text/text-line-height');
     Terra.validates.element('calculate line height');
   });
 
-  it('displays with Word Wrapped', () => {
+  it('should display with word wrapped', () => {
     browser.url('/raw/tests/terra-text/text/text-word-wrapped');
-    Terra.validates.element('word wrapped');
+    Terra.validates.element('word wrapped text');
   });
 });