+++ packages/terra-divider/CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Unreleased
 
 * Changed
+  * Removed unnecessary `wdio.conf.js` file
+  * Upgrade tests to terra-functional-testing and WebDriverIO v6
   * Updated Jest command
 
 ## 3.34.0 - (January 5, 2021)